feat(UsersDelete) - added a new table and models and a controller [Finishes#167729804]

+++ src/controllers/commentController.js
@@ -56,7 +56,7 @@ class CommentController {
             const { id } = req.params;
             
            
-            const comment = await Comment.findOne({where: {id}});
+            const comment = await Comment.findOne({ where: { userId: id}});
            
             if (!comment) {
                 return res.status(404).json({
@@ -64,7 +64,7 @@ class CommentController {
                     error: 'Comment Not Found',
                 });
             }
-            await Comment.destroy({ where: { id }});
+            await Comment.destroy({ where: { userId: id }});
             return res.status(200).json({
                 status: "success",
                 message: "Your comment has been Sucessfully Deleted"
